Implement LATERAL subqueries
For some context, see the dicussion on #460.

I define `lateral` in `Opaleye.Internal.QueryArr` so that we can use it to define `ArrowApply` and `Monad` instances for `QueryArr`.

I re-export it from `Query.Lateral`, along with `laterally` and `bilaterally`, which also came up in the aforementioned discussion.

This would fix #359.