distinctOn an empty set of columns should be `limit 1`, not a no-op
I understand why this is made a no-op, because an empty list of columns in a `DISTINCT ON` query is not syntactically valid in SQL. However, our Haskell types don't forbid it. So the question is, semantically, what does this mean? And I think it has to be an unconditional `limit 1`.